Add 'pull_date' Field to Card_Stocks Table
The data type of the 'pull_date' field in the 'Part_Of' table has been
changed to DATETIME from TIMESTAMP (TIMESTAMP doesn't play well with
others) and a corresponding 'pull_date' field was added to the
'Card_Stocks' table. These fields ARE NOT LINKED BY A FOREIGN KEY
because MySQL was apparently against allowing that.

ERRORS:
    ERROR 1215 (HY000): Cannot add foreign key constraint
    --> Triggered By:
                       ALTER TABLE 'Card_Stocks'
		       ADD CONSTRAINT fk_card_stocks_date FOREIGN KEY (pull_date)
		       REFERENCES Part_Of(pull_date);